重构：移除了 yml 配置文件，现在可以通过 .db.json 定制数据库连接。
