+++ .vscode/tasks.json
@@ -4,12 +4,12 @@
         {
             "label": "Prepare", // 用于准备测试环境。
             "type": "shell",
-            "command": "GODEBUG=tls10default=1 go run ./internal/testcmd/main.go -a PREPARE -c test_conf.yml",
+            "command": "GODEBUG=tls10default=1 go run ./internal/testcmd/main.go -a PREPARE",
         },
         {
             "label": "Clean", // 用于销毁测试环境。
             "type": "shell",
-            "command": "GODEBUG=tls10default=1 go run ./internal/testcmd/main.go -a CLEAN -c test_conf.yml",
+            "command": "GODEBUG=tls10default=1 go run ./internal/testcmd/main.go -a CLEAN",
         },
         {
             "label": "Test", // 测试本库。
