+++ internal/testcmd/main.go
@@ -17,37 +17,34 @@ var configFile string // 配置文件路径。
 
 func init() {
 	flag.StringVar(&action, "a", "", "execute action.")
-	flag.StringVar(&configFile, "c", "test_conf.yml", "conf file path")
+	flag.StringVar(&configFile, "c", ".", "conf file path")
 	flag.Parse()
 }
 
 // 这个 cmd 用于准备或销毁测试环境数据。
 func main() {
-	conf, err := testenv.LoadTestConfig(configFile)
-	if err != nil {
-		log.Fatal(err)
-	}
+	testenv.TryInitConfig(configFile)
 
 	switch action {
 	// 准备测试表。
 	case "PREPARE":
-		Prepare(conf)
+		Prepare()
 
 	// 销毁测试表。
 	case "CLEAN":
-		Clean(conf)
+		Clean()
 
 	default:
 		log.Fatal("undefined action:" + action)
 	}
 }
 
-func Prepare(testConf testenv.TestConf) {
+func Prepare() {
 	var errgroup errgroup.Group
 
 	// 初始化SqlServer测试表。
 	errgroup.Go(func() error {
-		db, err := getDb(mssql.DriverName, testConf.SqlServer)
+		db, err := getDb(mssql.DriverName, testenv.TestConf.SqlServer)
 		if err != nil {
 			return err
 		}
@@ -56,7 +53,7 @@ func Prepare(testConf testenv.TestConf) {
 
 	// 初始化 MySql 测试表。
 	errgroup.Go(func() error {
-		db, err := getDb(mysql.DriverName, testConf.MySql)
+		db, err := getDb(mysql.DriverName, testenv.TestConf.Mysql)
 		if err != nil {
 			return err
 		}
@@ -69,12 +66,12 @@ func Prepare(testConf testenv.TestConf) {
 	fmt.Println("Prepared.")
 }
 
-func Clean(testConf testenv.TestConf) {
+func Clean() {
 	var errgroup errgroup.Group
 
 	// 销毁 SqlServer 测试表。
 	errgroup.Go(func() error {
-		db, err := getDb(mssql.DriverName, testConf.SqlServer)
+		db, err := getDb(mssql.DriverName, testenv.TestConf.SqlServer)
 		if err != nil {
 			return err
 		}
@@ -83,7 +80,7 @@ func Clean(testConf testenv.TestConf) {
 
 	// 销毁 MySql 测试表。
 	errgroup.Go(func() error {
-		db, err := getDb(mysql.DriverName, testConf.MySql)
+		db, err := getDb(mysql.DriverName, testenv.TestConf.Mysql)
 		if err != nil {
 			return err
 		}
