+++ mssql/mssql_db_client_test.go
@@ -1,7 +1,6 @@
-package mssql
+package mssql_test
 
 import (
-	"database/sql"
 	"reflect"
 	"strings"
 	"testing"
@@ -11,31 +10,24 @@ import (
 
 	"github.com/bunnier/sqlmer"
 	"github.com/bunnier/sqlmer/internal/testenv"
+	"github.com/bunnier/sqlmer/mssql"
 )
 
-// 初始化测试配置。
-var testConf testenv.TestConf = testenv.MustLoadTestConfig("../test_conf.yml")
-
-// 用于获取一个 SqlServer 数据库的 DbClient 对象。
-func getMsSqlDbClient() (sqlmer.DbClient, error) {
-	return NewMsSqlDbClient(
-		testConf.SqlServer,
-		sqlmer.WithConnTimeout(time.Second*15),
-		sqlmer.WithExecTimeout(time.Second*15),
-	)
+func init() {
+	testenv.TryInitConfig("..")
 }
 
 func Test_NewMsSqlDbClient(t *testing.T) {
-	dbClient, err := getMsSqlDbClient()
+	dbClient, err := testenv.NewSqlServerClient()
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	if dbClient.Dsn() != testConf.SqlServer {
-		t.Errorf("mssqlDbClient.Dsn() connString = %v, wantConnString %v", dbClient.Dsn(), testConf.SqlServer)
+	if dbClient.Dsn() != testenv.TestConf.SqlServer {
+		t.Errorf("mssqlDbClient.Dsn() connString = %v, wantConnString %v", dbClient.Dsn(), testenv.TestConf.SqlServer)
 	}
 
-	if dbClient, err = NewMsSqlDbClient("test",
+	if dbClient, err = mssql.NewMsSqlDbClient("test",
 		sqlmer.WithConnTimeout(time.Second*15),
 		sqlmer.WithExecTimeout(time.Second*15),
 		sqlmer.WithPingCheck(true)); err == nil {
@@ -43,65 +35,8 @@ func Test_NewMsSqlDbClient(t *testing.T) {
 	}
 }
 
-func Test_bindMsSqlArgs(t *testing.T) {
-	testCases := []struct {
-		name      string
-		oriSql    string
-		args      []any
-		wantSql   string
-		wantParam []any
-		wantErr   error
-	}{
-		{
-			"map",
-			"SELECT * FROM go_TypeTest WHERE Id=@id",
-			[]any{
-				map[string]any{
-					"id": 1,
-				},
-			},
-			"SELECT * FROM go_TypeTest WHERE Id=@id",
-			[]any{sql.Named("id", 1)},
-			nil,
-		},
-		{
-			"index",
-			"SELECT * FROM go_TypeTest WHERE Id=@p1 OR Id=@p2",
-			[]any{1, 2},
-			"SELECT * FROM go_TypeTest WHERE Id=@p1 OR Id=@p2",
-			[]any{1, 2},
-			nil,
-		},
-	}
-
-	for _, tt := range testCases {
-		t.Run(tt.name, func(t *testing.T) {
-			fixedSql, args, err := bindArgs(tt.oriSql, tt.args...)
-
-			if tt.wantErr != nil {
-				if !errors.As(err, &tt.wantErr) {
-					t.Errorf("mssqlDbClient.bindMsSqlArgs() error = %v, wantErr %v", err, tt.wantErr)
-					return
-				}
-			} else {
-				if err != nil {
-					t.Error(err)
-					return
-				}
-				if fixedSql != tt.wantSql {
-					t.Errorf("mssqlDbClient.bindMsSqlArgs() sql = %v, wantSql %v", fixedSql, tt.wantSql)
-				}
-
-				if !reflect.DeepEqual(args, tt.wantParam) {
-					t.Errorf("mssqlDbClient.bindMsSqlArgs() args = %v, wantParam %v", args, tt.wantParam)
-				}
-			}
-		})
-	}
-}
-
 func Test_internalDbClient_Scalar(t *testing.T) {
-	mssqlClient, err := getMsSqlDbClient()
+	mssqlClient, err := testenv.NewSqlServerClient()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -142,7 +77,7 @@ func Test_internalDbClient_Scalar(t *testing.T) {
 }
 
 func Test_internalDbClient_Execute(t *testing.T) {
-	mssqlClient, err := getMsSqlDbClient()
+	mssqlClient, err := testenv.NewSqlServerClient()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -194,7 +129,7 @@ func Test_internalDbClient_Execute(t *testing.T) {
 }
 
 func Test_internalDbClient_Exists(t *testing.T) {
-	mssqlClient, err := getMsSqlDbClient()
+	mssqlClient, err := testenv.NewSqlServerClient()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -245,7 +180,7 @@ func Test_internalDbClient_Exists(t *testing.T) {
 }
 
 func Test_internalDbClient_Get(t *testing.T) {
-	mssqlClient, err := getMsSqlDbClient()
+	mssqlClient, err := testenv.NewSqlServerClient()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -380,7 +315,7 @@ func Test_internalDbClient_Get(t *testing.T) {
 }
 
 func Test_internalDbClient_Rows(t *testing.T) {
-	mssqlClient, err := getMsSqlDbClient()
+	mssqlClient, err := testenv.NewSqlServerClient()
 	if err != nil {
 		t.Fatal(err)
 	}
